Extract GitHub interfaces, make it a true plugin
