+++ src/main/java/nl/tudelft/ewi/sorcerers/github/GitHubCompareService.java
@@ -4,11 +4,13 @@ import java.io.IOException;
 
 import javax.inject.Inject;
 
+import nl.tudelft.ewi.sorcerers.model.CompareService;
+
 import org.eclipse.egit.github.core.RepositoryCommitCompare;
 import org.eclipse.egit.github.core.RepositoryId;
 import org.eclipse.egit.github.core.service.CommitService;
 
-public class GitHubCompareService {
+public class GitHubCompareService implements CompareService {
 	private CommitService commitService;
 	
 	@Inject
@@ -16,6 +18,7 @@ public class GitHubCompareService {
 		this.commitService = commitService;
 	}
 	
+	@Override
 	public String getMergeBase(String repo, String base, String head)
 			throws IOException {
 		RepositoryCommitCompare compare = this.commitService.compare(RepositoryId.createFromId(repo), base, head);
