+++ src/main/java/nl/tudelft/ewi/sorcerers/model/CompareService.java
@@ -1,0 +1,10 @@
+package nl.tudelft.ewi.sorcerers.model;
+
+import java.io.IOException;
+
+public interface CompareService {
+
+	public abstract String getMergeBase(String repo, String base, String head)
+			throws IOException;
+
+}
